Delete temporary files when closing Fiji
Temporary files are created when starting Blender. They will accumulate in
the temp directory. This changes makes sure that the files are deleted when
closing Fiji. So that the user now longer needs to clean them up manually.
This comes with the risk that Blender tries to access these files after
Fiji was closed, but I assume that's not a problem, because the files are
only needed when Blender starts.