Fix read_csv.py
The edges parameter that was passed to the "add_mesh" method where wrong.
"edges" should be a list of pairs of indices into the "vertices" list.
However previously the "edges" where the pool indices used by mastodon
core. However the pool index matches the index in the list of vertices
only if no spots have been deleted. This wrong indices would cause
Blender to crash with a seg fault on some occasions.

This commit also passes numpy arrays instead of lists and avoids loops
where possible.